Utilize uint8 for MVV/LVA scores (0x10's place being victim, 0x01's place the attacker) rather than uint16 (with 100's and 1's). Doesn't actually save space in the hash table on x86_64, presumably due to struct padding.

+++ src/board.h
@@ -83,8 +83,8 @@ typedef struct board_s
 	uint8 trades;
 
 	// move ordering:
-	uint16 victim [33];
-	uint16 attack [33];
+	uint8 victim [33];
+	uint8 attack [33];
 } board;
 
 extern board *curboard;
